WIP: gracefully fail when listing reports on unavailable data points.
